SLM-64: Implemented nunjucks filter for Review Recipients table
