+++ server/filters/formatDateForResultsPageFilter.ts
@@ -1,0 +1,6 @@
+import moment from 'moment'
+
+export default function formatDateForResultsPageFilter(value: string): string {
+  const dateTime = moment(value)
+  return dateTime ? dateTime.format('D MMMM YYYY') : null
+}
