+++ server/views/pages/barcode/review-recipients.njk
@@ -30,7 +30,7 @@
             { text: '' },
             { text: '' }
           ],
-          rows: recipients
+          rows: recipients | recipientTableRows
         }) }}
 
         <p class="govuk-!-margin-top-6 govuk-!-margin-bottom-6">
